+++ comicbox/comic_archive.py
@@ -1,16 +1,24 @@
 """
-Comic archive.
+Comic Archive.
 
 Reads and writes metadata via the included metadata package.
 Reads data using libarchive via archi.
 """
-
+import tarfile
 import zipfile
 
+from functools import wraps
+from io import BytesIO
 from logging import getLogger
 from pathlib import Path
+from typing import Callable
+from typing import Optional
+from typing import Union
 
 import rarfile
+
+from confuse import AttrDict
+from PIL import Image
 
 from comicbox.config import get_config
 from comicbox.exceptions import UnsupportedArchiveTypeError
@@ -31,7 +39,18 @@ init_logging()
 LOG = getLogger(__name__)
 
 
-class ComicArchive(object):
+def _archive_close(f):
+    @wraps(f)
+    def wrapper(self, *args, **kwargs):
+        result = f(self, *args, **kwargs)
+        if self._closefd:
+            self.close()
+        return result
+
+    return wrapper
+
+
+class ComicArchive:
     """
     Represent a comic archive.
 
@@ -40,167 +59,255 @@ class ComicArchive(object):
 
     PARSER_CLASSES = (ComicInfoXml, ComicBookInfo, CoMet)
     FILENAMES = set((CoMet.FILENAME, ComicInfoXml.FILENAME))
+    _PAGE_KEYS = set(("page_count", "cover_image"))
+    _PAGES_KEYS = set(("pages",)) | _PAGE_KEYS
+    _RAW_CBI_KEY = "ComicBookInfo Archive Comment"
+    _RAW_FILENAME_KEY = "Filename"
 
-    def __init__(self, path, metadata=None, config=None):
-        """Initialize the archive with a path to the archive."""
+    def __init__(
+        self,
+        path: Union[Path, str],
+        config: Optional[AttrDict] = None,
+        metadata: Optional[dict] = None,
+        closefd: bool = True,
+    ):
+        """Initialize the archive with a path to the archive.
+
+        path: the path to the comic archive
+        config: a confuse AttrDict. If None, ComicArchive generates its own from the
+            environment.
+        metadata: a comicbox metadata dict to use instead of gathering the metadata
+            from the path.
+        closefd: whether or not to close the comic archive after every public method
+            call or leave it open. If set to False, you should call
+            ComicArchive.close() when done with the comic archive.
+        """
+        self._path: Path = Path(path)
         if config is None:
             config = get_config()
-        self.config = config
-        self.set_path(path)
-        self.metadata = ComicBaseMetadata(metadata=metadata)
-        self.raw = {}
-        self.cover_image_data = None
-        self._parse_metadata()
+        self._config: AttrDict = config
+        self._set_archive_cls()
+        self._archive: Union[
+            zipfile.ZipFile, rarfile.RarFile, tarfile.TarFile, None
+        ] = None
+        self._metadata: ComicBaseMetadata = ComicBaseMetadata(metadata=metadata)
+        self._closefd: bool = closefd
+        self._raw: dict = {}
 
-    def set_path(self, path):
+    def _set_archive_cls(self):
         """Set the path and determine the archive type."""
-        self._path = Path(path)
+        self._archive_cls: Callable
         if zipfile.is_zipfile(self._path):
-            self.archive_cls = zipfile.ZipFile
+            self._archive_cls = zipfile.ZipFile
         elif rarfile.is_rarfile(self._path):
-            self.archive_cls = rarfile.RarFile
+            self._archive_cls = rarfile.RarFile
+        elif tarfile.is_tarfile(self._path):
+            self._archive_cls = tarfile.open
         else:
             raise UnsupportedArchiveTypeError(f"Unsupported archive type: {self._path}")
 
-    def _get_archive(self, mode="r"):
-        return self.archive_cls(self._path, mode=mode)  # type: ignore
+    def _get_archive(self):
+        """Set archive instance open for reading."""
+        if not self._archive:
+            self._archive = self._archive_cls(self._path)
+        return self._archive
 
-    def get_path(self):
-        """Get the path for the archive."""
-        return self._path
+    def _archive_namelist(self):
+        """Get list of files in the archive."""
+        archive = self._get_archive()
+        if isinstance(archive, tarfile.TarFile):
+            namelist = archive.getnames()
+        else:
+            namelist = archive.namelist()
+        return sorted(namelist)
 
-    def namelist(self):
-        """Get the archive file namelist."""
-        with self._get_archive() as archive:
-            return sorted(archive.namelist())
+    def _archive_infolist(self):
+        """Get info list of members from the archive."""
+        archive = self._get_archive()
+        if isinstance(archive, tarfile.TarFile):
+            infolist = archive.getmembers()
+        else:
+            infolist = archive.infolist()
+        if self._archive_cls == tarfile.open:
+            fn_attr = "name"
+        else:
+            fn_attr = "filename"
+        infolist = sorted(infolist, key=lambda i: getattr(i, fn_attr))
+        return infolist, fn_attr
 
-    def _parse_xml_metadata(self, fn, xml_parser_cls, flag, archive):
-        """Run the correct parser for the xml file."""
-        md_filename = str(xml_parser_cls.FILENAME)
-        if Path(fn).name.lower() != md_filename.lower() or not flag:
+    def _archive_readfile(self, filename):
+        """Read an archive file to memory."""
+        archive = self._get_archive()
+        if isinstance(archive, tarfile.TarFile):
+            file_obj = archive.extractfile(filename)
+            if file_obj:
+                return file_obj.read()
+        else:
+            return archive.read(filename)  # type: ignore
+
+    def _get_raw_files_metadata(self):
+        """Get raw metadata from files in the archive."""
+
+        # create parser_classes_dict
+        all_parser_classes = {
+            CoMet: self._config.comet,
+            ComicInfoXml: self._config.comicinfoxml,
+        }
+        parser_classes = {}
+        for parser_class, flag in all_parser_classes.items():
+            if flag and not self._raw.get(parser_class.FILENAME):
+                parser_classes[parser_class.FILENAME] = parser_class
+
+        # search filenames for metadata files and read.
+        result_parsers = {}
+        for fn in self._archive_namelist():
+            lower_name = Path(fn).name.lower()
+            parser_class = parser_classes.get(lower_name)
+            if parser_class and not self._raw.get(parser_class.FILENAME):
+                data = self._archive_readfile(fn)
+                self._raw[parser_class.FILENAME] = data
+                result_parsers[parser_class] = data
+                del parser_classes[parser_class.FILENAME]
+                if not parser_classes:
+                    break
+        return result_parsers
+
+    def _parse_files_metadata(self):
+        """Run the correct parser for the each archive file's metadata."""
+        result_parsers = self._get_raw_files_metadata()
+        files_md = {}
+        for parser_class, data in result_parsers.items():
+            parser = parser_class(string=data)
+            files_md[parser_class] = parser.metadata
+        return files_md
+
+    def _get_raw_archive_comment(self):
+        """Get the comment field from an archive."""
+        if (
+            self._archive_cls != tarfile.open
+            and self._config.comicbookinfo
+            and not self._raw.get(self._RAW_CBI_KEY)
+        ):
+            comment = self._get_archive().comment  # type: ignore
+            if isinstance(comment, bytes):
+                comment = comment.decode()
+            if comment:
+                self._raw[self._RAW_CBI_KEY] = comment
+        return self._raw.get(self._RAW_CBI_KEY)
+
+    def _parse_metadata_comments(self):
+        """Parse the metadata comments with CBI."""
+        data = self._get_raw_archive_comment()
+        if not data:
             return {}
-
-        data = archive.read(fn)
-        if self.config.raw:
-            self.raw[md_filename] = data
-        parser = xml_parser_cls(string=data)
+        parser = ComicBookInfo(string=data)
         return parser.metadata
 
-    def _parse_metadata_entries(self, archive):
-        """Get the filenames and file based metadata."""
-        cix_md = {}
-        comet_md = {}
-        for fn in sorted(archive.namelist()):
-            cix_md.update(
-                self._parse_xml_metadata(
-                    fn, ComicInfoXml, self.config.comicinfoxml, archive
-                )
-            )
-            comet_md.update(
-                self._parse_xml_metadata(fn, CoMet, self.config.comet, archive)
-            )
-        return cix_md, comet_md
-
-    def _get_archive_comment(self, archive):
-        """Get the comment field from an archive."""
-        if not archive:
-            archive = self._get_archive()
-        comment = archive.comment
-        if isinstance(comment, bytes):
-            comment = comment.decode()
-        return comment
-
-    def _parse_metadata_comments(self, archive):
-        if not self.config.comicbookinfo:
-            return {}
-        comment = self._get_archive_comment(archive)
-        parser = ComicBookInfo(string=comment)
-        cbi_md = parser.metadata
-        if self.config.raw:
-            self.raw["ComicBookInfo Archive Comment"] = comment
-        return cbi_md
+    def _get_raw_filename(self):
+        """Get the filename form the path."""
+        if self._config.filename and not self._raw.get(self._RAW_FILENAME_KEY):
+            data = self._path.name
+            self._raw[self._RAW_FILENAME_KEY] = data
+        return self._raw.get(self._RAW_FILENAME_KEY)
 
     def _parse_metadata_filename(self):
-        if not self.config.filename:
+        """Parse metadata from the filename."""
+        data = self._get_raw_filename()
+        if data is None:
             return {}
-        parser = FilenameMetadata(path=self._path)
-        if self.config.raw:
-            self.raw["Filename"] = self._path.name
+        parser = FilenameMetadata(path=data)
         return parser.metadata
 
-    def _parse_metadata(self):
+    def _ensure_page_metadata(self):
+        """Ensure page metadata exists."""
         if (
-            self.metadata.metadata
-            and not self.config.metadata
-            and not self.config.cover
+            not self._PAGE_KEYS.issubset(self._metadata.metadata)
+            or self._metadata.get_num_pages() is None
         ):
-            return
+            namelist = self._archive_namelist()
+            self._metadata.set_page_metadata(namelist)
+
+    def _parse_metadata(self):
+        """Parse all enabled metadata."""
         md_list = []
-        if self.config.filename:
-            filename_md = self._parse_metadata_filename()
+        filename_md = self._parse_metadata_filename()
+        if filename_md:
             md_list += [filename_md]
-            self.metadata.synthesize_metadata(md_list)
+        files_md = self._parse_files_metadata()
+        comet_md = files_md.get(CoMet)
+        if comet_md:
+            md_list += [comet_md]
+        cbi_md = self._parse_metadata_comments()
+        if cbi_md:
+            md_list += [cbi_md]
+        cix_md = files_md.get(ComicInfoXml)
+        if cix_md:
+            md_list += [cix_md]
+        # order of the md list is very important, lowest to highest
+        # precedence.
+        self._metadata.synthesize_metadata(md_list)
+        self._ensure_page_metadata()
 
-        if not (
-            self.config.comicinfoxml
-            or self.config.comicbookinfo
-            or self.config.comet
-            or self.config.cover
-        ):
-            return
-        with self._get_archive() as archive:
-            if (
-                self.config.comicinfoxml
-                or self.config.comicbookinfo
-                or self.config.comet
-            ):
-                cbi_md = self._parse_metadata_comments(archive)
-                cix_md, comet_md = self._parse_metadata_entries(archive)
-                # order of the md list is very important, lowest to highest
-                # precedence.
-                md_list += [comet_md, cbi_md, cix_md]
-                self.metadata.synthesize_metadata(md_list)
-                self._ensure_page_metadata(archive)
-            if self.config.cover:
-                self.cover_image_data = self._get_cover_image(archive)
+    def _set_raw_metadata(self):
+        """Set only the raw metadata."""
+        self._get_raw_filename()
+        self._get_raw_archive_comment()
+        self._get_raw_files_metadata()
 
+    def _write_cbi_comment(self, parser):
+        """Write a cbi comment to an archive."""
+        if self._archive_cls == tarfile.open:
+            raise ValueError("Cannot write ComicBookInfo comments to cbt tarfile.")
+        self.close()
+        with self._archive_cls(self._path, "a") as append_archive:
+            comment = parser.to_string().encode()
+            append_archive.comment = comment  # type: ignore
+
+    def close(self):
+        """Close the open archive."""
+        try:
+            if self._archive:
+                self._archive.close()
+        except Exception:
+            pass
+        self._archive = None
+
+    @_archive_close
     def get_num_pages(self):
         """Return the number of pages."""
-        return self.metadata.get_num_pages()
+        self._ensure_page_metadata()
+        return self._metadata.get_num_pages()
 
-    def _ensure_page_metadata(self, archive):
-        """Ensure page metadata exists."""
-        compute = False
-        for key in ("page_count", "cover_image"):
-            if not self.metadata.metadata.get(key):
-                compute = True
-        if compute:
-            namelist = sorted(archive.namelist())
-            self.metadata.compute_page_metadata(namelist)
-
+    @_archive_close
     def get_pages(self, page_from):
         """Generate all pages starting with page number."""
-        with self._get_archive() as archive:
-            self._ensure_page_metadata(archive)
-            pagenames = self.metadata.get_pagenames_from(page_from)
+        self._ensure_page_metadata()
+        pagenames = self._metadata.get_pagenames_from(page_from)
+        if pagenames:
             for pagename in pagenames:
-                with archive.open(pagename) as page:
-                    yield page.read()
+                yield self._archive_readfile(pagename)
 
+    @_archive_close
     def get_page_by_filename(self, filename):
         """Return data for a single page by filename."""
-        with self._get_archive() as archive:
-            with archive.open(filename) as page:
-                return page.read()
+        data = self._archive_readfile(filename)
+        return data
 
+    @_archive_close
     def get_page_by_index(self, index):
         """Get the page data by index."""
-        with self._get_archive() as archive:
-            self._ensure_page_metadata(archive)
-            filename = self.metadata.get_pagename(index)
-            with archive.open(filename) as page:
-                return page.read()
+        self._ensure_page_metadata()
+        filename = self._metadata.get_pagename(index)
+        data = self._archive_readfile(filename)
+        return data
 
+    @_archive_close
+    def get_page_by_index_as_pil(self, index):
+        """ "Return page as pil image."""
+        data = self.get_page_by_index(index)
+        return Image.open(BytesIO(data))
+
+    @_archive_close
     def extract_pages(self, page_from, root_path="."):
         """Extract pages from archive and write to a path."""
         root_path = Path(root_path)
@@ -209,62 +316,66 @@ class ComicArchive(object):
                 f"Must extract pages to a directory. {str(root_path)} "
                 "is not a directory"
             )
-        with self._get_archive() as archive:
-            self._ensure_page_metadata(archive)
-            filenames = self.metadata.get_pagenames_from(page_from)
-            for fn in filenames:
-                with archive.open(fn) as page:
-                    if self.config.dry_run:
-                        LOG.info(f"Not extracting page from {self._path}: {fn}")
-                        continue
-                    full_path = Path(root_path) / Path(fn).name
-                    with full_path.open("wb") as page_file:
-                        page_file.write(page.read())
+        self._ensure_page_metadata()
+        pagenames = self._metadata.get_pagenames_from(page_from)
+        if pagenames:
+            for fn in pagenames:
+                if self._config.dry_run:
+                    LOG.info(f"Not extracting page from {self._path}: {fn}")
+                    continue
+                full_path = Path(root_path) / Path(fn).name
+                with full_path.open("wb") as page_file:
+                    page_file.write(self._archive_readfile(fn))
 
+    @_archive_close
     def extract_cover_as(self, path):
         """Extract the cover image to a destination file."""
-        with self._get_archive() as archive:
-            self._ensure_page_metadata(archive)
-            cover_fn = self.metadata.get_cover_page_filename()
-            if not cover_fn:
-                return
-            if self.config.dry_run:
-                LOG.info(f"Not extracting cover from {self._path}: {cover_fn}")
-                return
-            output_path = Path(path)
-            if output_path.is_dir():
-                output_path = output_path / Path(cover_fn).name
-            with archive.open(cover_fn) as page:
-                with output_path.open("wb") as cover_file:
-                    cover_file.write(page.read())
+        self._ensure_page_metadata()
+        cover_fn = self._metadata.get_cover_page_filename()
+        if not cover_fn:
+            return
+        if self._config.dry_run:
+            LOG.info(f"Not extracting cover from {self._path}: {cover_fn}")
+            return
+        output_path = Path(path)
+        if output_path.is_dir():
+            output_path = output_path / Path(cover_fn).name
+        with output_path.open("wb") as cover_file:
+            cover_file.write(self._archive_readfile(cover_fn))
 
-    def _get_cover_image(self, archive):
+    @_archive_close
+    def get_cover_image(self):
         """Return cover image data."""
-        self._ensure_page_metadata(archive)
-        cover_fn = self.metadata.get_cover_page_filename()
+        self._ensure_page_metadata()
+        cover_fn = self._metadata.get_cover_page_filename()
         if not cover_fn:
             return
         try:
-            data = archive.read(cover_fn)
+            data = self._archive_readfile(cover_fn)
         except Exception as exc:
             LOG.error(f"{self._path} reading cover: {cover_fn}")
             raise exc
         return data
 
-    def get_cover_image(self):
-        """Get the cover image."""
-        if not self.cover_image_data:
-            with self._get_archive() as archive:
-                self.cover_image_data = self._get_cover_image(archive)
-        return self.cover_image_data
+    @_archive_close
+    def get_cover_image_as_pil(self):
+        """Get the cover image in PIL form."""
+        image = self.get_cover_image()
+        return Image.open(BytesIO(image))
 
+    @_archive_close
     def get_metadata(self):
         """Return the metadata from the archive."""
-        return self.metadata.metadata
+        if not self._metadata.metadata or bool(
+            set(self._metadata.metadata.keys()) - self._PAGES_KEYS
+        ):
+            self._parse_metadata()
+        return self._metadata.metadata
 
+    @_archive_close
     def recompress(self, filename=None, data=None):
         """Recompress the archive optionally replacing a file."""
-        if self.config.dry_run:
+        if self._config.dry_run:
             LOG.info(f"Not recompressing: {self._path}")
             return
 
@@ -273,45 +384,47 @@ class ComicArchive(object):
             raise ValueError(f"{new_path} already exists.")
 
         tmp_path = self._path.with_suffix(RECOMPRESS_SUFFIX)
-        with self._get_archive() as archive:
-            if self.config.delete_tags:
-                comment = b""
-            else:
-                comment = archive.comment
+        comment = b""
+        if self._archive_cls != tarfile.open:
+            if not self._config.delete_tags:
+                comment = self._get_archive().comment  # type: ignore
             if isinstance(comment, str):
                 comment = comment.encode()
-            with zipfile.ZipFile(
-                tmp_path, "w", compression=zipfile.ZIP_DEFLATED, compresslevel=9
-            ) as zf:
-                skipnames = set()
-                if filename:
-                    skipnames.add(filename)
-                if self.config.delete_tags:
-                    skipnames.add(self.FILENAMES)
-                for info in sorted(archive.infolist(), key=lambda i: i.filename):
-                    if info.filename.lower() in skipnames:
-                        continue
-                    if IMAGE_EXT_RE.search(info.filename) is None:
-                        compress = zipfile.ZIP_DEFLATED
-                    else:
-                        # images usually end up slightly larger with
-                        # zip compression
-                        compress = zipfile.ZIP_STORED
-                    zf.writestr(
-                        info.filename,
-                        archive.read(info),
-                        compress_type=compress,
-                        compresslevel=9,
-                    )
-                if filename and data:
-                    zf.writestr(filename, data)
-                if comment:
-                    zf.comment = comment
+
+        with zipfile.ZipFile(
+            tmp_path, "w", compression=zipfile.ZIP_DEFLATED, compresslevel=9
+        ) as zf:
+            skipnames = set()
+            if filename:
+                skipnames.add(filename)
+            if self._config.delete_tags:
+                skipnames.add(self.FILENAMES)
+            infolist, fn_attr = self._archive_infolist()
+            for info in infolist:
+                fn = getattr(info, fn_attr)
+                if fn.lower() in skipnames:
+                    continue
+                if IMAGE_EXT_RE.search(fn) is None:
+                    compress = zipfile.ZIP_DEFLATED
+                else:
+                    # images usually end up slightly larger with
+                    # zip compression
+                    compress = zipfile.ZIP_STORED
+                zf.writestr(
+                    fn,
+                    self._archive_readfile(fn),
+                    compress_type=compress,
+                    compresslevel=9,
+                )
+            if filename and data:
+                zf.writestr(filename, data)
+            if comment:
+                zf.comment = comment
 
         old_path = self._path
         tmp_path.replace(new_path)
         self._path = new_path
-        if self.config.delete_rar:
+        if self._config.delete_rar:
             LOG.info(f"converted to: {new_path}")
             if new_path.is_file():
                 old_path.unlink()
@@ -319,7 +432,7 @@ class ComicArchive(object):
 
     def write_metadata(self, md_class, recompute_page_sizes=True):
         """Write metadata using the supplied parser class."""
-        if self.config.dry_run:
+        if self._config.dry_run:
             LOG.info(f"Not writing metadata for: {self._path}")
             return
         parser = md_class(metadata=self.get_metadata())
@@ -328,9 +441,7 @@ class ComicArchive(object):
         if isinstance(parser, (ComicXml, CoMet)):
             self.recompress(parser.FILENAME, parser.to_string())
         elif isinstance(parser, ComicBookInfo):
-            with self._get_archive("a") as archive:
-                comment = parser.to_string().encode()
-                archive.comment = comment
+            self._write_cbi_comment(parser)
         else:
             raise ValueError(f"Unsupported metadata writer {md_class}")
 
@@ -355,12 +466,12 @@ class ComicArchive(object):
             except (ParseError, JSONDecodeError):
                 pass
         if success_class and md:
-            self.metadata.metadata = md.metadata
+            self._metadata.metadata = md.metadata
             self.write_metadata(success_class)
 
     def export_files(self):
         """Export metadata to all supported file formats."""
-        if self.config.dry_run:
+        if self._config.dry_run:
             LOG.info("Not exporting files.")
             return
         for parser_cls in self.PARSER_CLASSES:
@@ -368,31 +479,39 @@ class ComicArchive(object):
             path = Path(str(parser_cls.FILENAME))
             md.to_file(path)
 
+    @_archive_close
     def compute_pages_tags(self):
         """Recompute the tag image sizes for ComicRack."""
-        with self._get_archive() as archive:
-            infolist = archive.infolist()
-        parser = ComicInfoXml(metadata=self.get_metadata())
+        infolist, _ = self._archive_infolist()
+        metadata = self.get_metadata()
+        parser = ComicInfoXml(metadata=metadata)
         parser.compute_pages_tags(infolist)
-        self.metadata.metadata["pages"] = parser.metadata.get("pages")
-
-    def compute_page_count(self):
-        """Compute the page count from images in the archive."""
-        self.metadata.compute_page_count()
+        self._metadata.metadata["pages"] = parser.metadata.get("pages")
 
     def rename_file(self):
         """Rename the archive."""
-        if self.config.dry_run:
+        if self._config.dry_run:
             LOG.info(f"Not reaming file: {self._path}")
             return
 
-        car = FilenameMetadata(self.metadata)
+        car = FilenameMetadata(self._metadata)
         self._path = car.to_file(self._path)
 
+    @_archive_close
     def print_raw(self):
         """Print raw metadtata."""
-        for key, val in self.raw.items():
+        self._set_raw_metadata()
+        for key, val in self._raw.items():
             print("-" * 10, key, "-" * 10)
             if isinstance(val, bytes):
                 val = val.decode()
             print(val)
+
+    def get_path(self):
+        """Get the path for the archive."""
+        return self._path
+
+    @_archive_close
+    def namelist(self):
+        """Get the archive file namelist."""
+        return self._archive_namelist()