+++ tests/test_cbi.py
@@ -29,9 +29,10 @@ METADATA = {
     "language": "en",
     "country": "US",
     "title": "The Beginning",
-    "remainder": "cbi",
     "page_count": 36,
     "cover_image": "Captain Science 001/CaptainScience#1_01.jpg",
+    "ext": "cbr",
+    # "remainder": "cbi",
 }