+++ NEWS.md
@@ -1,5 +1,14 @@
 # 📰 News
 
+## v0.5.1
+
+- Features
+
+  - Methods for getting covers and pages as PIL Images.
+  - Lazy metadata and cover pulling. Removes `metadata` and `cover` options.
+  - closefd option leaves archive open for multiple method calls. Close manually with close().
+  - .cbt Tarfile comic archive support.
+
 ## v0.5.0
 
 - Features